chat-backend/requirements.txt
fastapi
uvicorn
pymongo
faiss-cpu
sentence-transformers
openai
python-dotenv
pydantic
transformers